+++ kurtosis-module/impl/participant_network/cl/nimbus/nimbus_launcher.go
@@ -55,7 +55,6 @@ const (
 
 	metricsPath = "/metrics"
 
-	grafanaDashboardConfigFilename = "nimbus.json"
 )
 var usedPorts = map[string]*services.PortSpec{
 	tcpDiscoveryPortID: services.NewPortSpec(discoveryPortNum, services.PortProtocol_TCP),
@@ -130,15 +129,13 @@ func (launcher NimbusLauncher) Launch(
 	metricsUrl := fmt.Sprintf("%v:%v", serviceCtx.GetPrivateIPAddress(), metricsPort.GetNumber())
 
 	nodeMetricsInfo := cl.NewCLNodeMetricsInfo(string(serviceId), metricsPath, metricsUrl)
-	clNodesMetricsInfo := []*cl.CLNodeMetricsInfo{nodeMetricsInfo}
-
-	metricsInfo := cl.NewCLMetricsInfo(grafanaDashboardConfigFilename, clNodesMetricsInfo)
+	nodesMetricsInfo := []*cl.CLNodeMetricsInfo{nodeMetricsInfo}
 
 	result := cl.NewCLClientContext(
 		nodeIdentity.ENR,
 		serviceCtx.GetPrivateIPAddress(),
 		httpPortNum,
-		metricsInfo,
+		nodesMetricsInfo,
 		restClient,
 	)